小修改

+++ login_web_practice/src/main/java/com/example/loginsystem/repository/PremiumUserRepository.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 import com.example.loginsystem.model.PremiumUser;
 
 @Repository
-public interface PremiumUserRepository extends JpaRepository<PremiumUser, Long> {
+public interface PremiumUserRepository extends JpaRepository<PremiumUser, Integer> {
     PremiumUser findByUsernameAndPassword(String username, String password);
     PremiumUser findByUsername(String username);
 }
