+++ login_web_practice/src/main/java/com/example/loginsystem/repository/UserRepository.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 import com.example.loginsystem.model.NormalUser;
 
 @Repository
-public interface UserRepository extends JpaRepository<NormalUser, Long> {
+public interface UserRepository extends JpaRepository<NormalUser, Integer> {
     NormalUser findByUsernameAndPassword(String username, String password);
     NormalUser findByUsername(String username);
 }
